pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Modular FastAPI authentication service with RBAC"
authors = [{ name = "Example", email = "example@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "pydantic-settings>=2.2",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]